fix: amended button and move to shared
