--- src/components/UI/Button/index.ts
@@ -1,4 +1,0 @@
-import { Button } from './Button';
-import { type TButton } from './typing';
-
-export { Button, type TButton };
