+++ src/shared/Button/typing.ts
@@ -1,7 +1,7 @@
+import { ButtonHTMLAttributes } from 'react';
+
 export type TButton = {
   label?: string;
   onClick?: () => void;
-  disabled?: boolean;
   type?: 'submit' | 'reset' | 'button';
-  children?: React.ReactNode;
-};
+} & ButtonHTMLAttributes<HTMLButtonElement>;
